Major Refactoring of All Code.
The mascot class was broken up in a couple of smaller specialist classes to handle animation and getting xml data. The calculation of speed and the application of addittional forces was also overhauled to a much more reliable system. Collision was fixed as well as the quality of rendering. The ReadMe file has also been rewritten to include an overview of the project.

A more detailed version of this is available in the current ReadMe.